slightly better examples
